+++ examples/src/json_cursor_examples.cpp
@@ -1,7 +1,7 @@
 // Copyright 2013 Daniel Parker
 // Distributed under Boost license
 
-#include <jsoncons/json_pull_reader.hpp>
+#include <jsoncons/json_cursor.hpp>
 #include <string>
 #include <sstream>
 
@@ -36,7 +36,7 @@ void reading_a_json_stream()
 {
     std::istringstream is(example);
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     for (; !reader.done(); reader.next())
     {
@@ -116,7 +116,7 @@ void filtering_a_json_stream()
     std::istringstream is(example);
 
     author_filter filter;
-    json_pull_reader reader(is, filter);
+    json_cursor reader(is, filter);
 
     for (; !reader.done(); reader.next())
     {