+++ tests/src/json_cursor_tests.cpp
@@ -6,7 +6,7 @@
 #endif
 #include <jsoncons/json.hpp>
 #include <jsoncons/json_encoder.hpp>
-#include <jsoncons/json_pull_reader.hpp>
+#include <jsoncons/json_cursor.hpp>
 #include <catch/catch.hpp>
 #include <sstream>
 #include <vector>
@@ -15,12 +15,12 @@
 
 using namespace jsoncons;
 
-TEST_CASE("json_pull_reader string_value test")
+TEST_CASE("json_cursor string_value test")
 {
     std::string s = R"("Tom")";
     std::istringstream is(s);
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     REQUIRE_FALSE(reader.done());
     CHECK(reader.current().event_type() == staj_event_type::string_value);
@@ -30,12 +30,12 @@ TEST_CASE("json_pull_reader string_value test")
     CHECK(reader.done());
 }
 
-TEST_CASE("json_pull_reader string_value as<int> test")
+TEST_CASE("json_cursor string_value as<int> test")
 {
     std::string s = R"("-100")";
     std::istringstream is(s);
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     REQUIRE_FALSE(reader.done());
     CHECK(reader.current().event_type() == staj_event_type::string_value);
@@ -44,12 +44,12 @@ TEST_CASE("json_pull_reader string_value as<int> test")
     CHECK(reader.done());
 }
 
-TEST_CASE("json_pull_reader string_value as<unsigned> test")
+TEST_CASE("json_cursor string_value as<unsigned> test")
 {
     std::string s = R"("100")";
     std::istringstream is(s);
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     REQUIRE_FALSE(reader.done());
     CHECK(reader.current().event_type() == staj_event_type::string_value);
@@ -59,12 +59,12 @@ TEST_CASE("json_pull_reader string_value as<unsigned> test")
     CHECK(reader.done());
 }
 
-TEST_CASE("json_pull_reader null_value test")
+TEST_CASE("json_cursor null_value test")
 {
     std::string s = "null";
     std::istringstream is(s);
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     REQUIRE_FALSE(reader.done());
     CHECK(reader.current().event_type() == staj_event_type::null_value);
@@ -72,12 +72,12 @@ TEST_CASE("json_pull_reader null_value test")
     CHECK(reader.done());
 }
 
-TEST_CASE("json_pull_reader bool_value test")
+TEST_CASE("json_cursor bool_value test")
 {
     std::string s = "false";
     std::istringstream is(s);
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     REQUIRE_FALSE(reader.done());
     CHECK(reader.current().event_type() == staj_event_type::bool_value);
@@ -85,12 +85,12 @@ TEST_CASE("json_pull_reader bool_value test")
     CHECK(reader.done());
 }
 
-TEST_CASE("json_pull_reader int64_value test")
+TEST_CASE("json_cursor int64_value test")
 {
     std::string s = "-100";
     std::istringstream is(s);
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     REQUIRE_FALSE(reader.done());
     CHECK(reader.current().event_type() == staj_event_type::int64_value);
@@ -99,12 +99,12 @@ TEST_CASE("json_pull_reader int64_value test")
     CHECK(reader.done());
 }
 
-TEST_CASE("json_pull_reader uint64_value test")
+TEST_CASE("json_cursor uint64_value test")
 {
     std::string s = "100";
     std::istringstream is(s);
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     REQUIRE_FALSE(reader.done());
     CHECK(reader.current().event_type() == staj_event_type::uint64_value);
@@ -114,12 +114,12 @@ TEST_CASE("json_pull_reader uint64_value test")
     CHECK(reader.done());
 }
 
-TEST_CASE("json_pull_reader string_value as bignum test")
+TEST_CASE("json_cursor string_value as bignum test")
 {
     std::string s = "-18446744073709551617";
     std::istringstream is("\""+s+"\"");
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     REQUIRE_FALSE(reader.done());
     CHECK(reader.current().event_type() == staj_event_type::string_value);
@@ -130,12 +130,12 @@ TEST_CASE("json_pull_reader string_value as bignum test")
     CHECK(reader.done());
 }
 
-TEST_CASE("json_pull_reader bigint value as bignum")
+TEST_CASE("json_cursor bigint value as bignum")
 {
     std::string s = "-18446744073709551617";
     std::istringstream is(s);
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     REQUIRE_FALSE(reader.done());
     CHECK(reader.current().event_type() == staj_event_type::string_value);
@@ -146,12 +146,12 @@ TEST_CASE("json_pull_reader bigint value as bignum")
     CHECK(reader.done());
 }
 
-TEST_CASE("json_pull_reader double_value test")
+TEST_CASE("json_cursor double_value test")
 {
     std::string s = "100.0";
     std::istringstream is(s);
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     REQUIRE_FALSE(reader.done());
     CHECK(reader.current().event_type() == staj_event_type::double_value);
@@ -159,7 +159,7 @@ TEST_CASE("json_pull_reader double_value test")
     CHECK(reader.done());
 }
 
-TEST_CASE("json_pull_reader array_value test")
+TEST_CASE("json_cursor array_value test")
 {
     std::string s = R"(
     [
@@ -185,7 +185,7 @@ TEST_CASE("json_pull_reader array_value test")
 
     std::istringstream is(s);
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     REQUIRE_FALSE(reader.done());
     CHECK(reader.current().event_type() == staj_event_type::begin_array);
@@ -286,7 +286,7 @@ TEST_CASE("json_pull_reader array_value test")
     CHECK(reader.done());
 }
 
-TEST_CASE("json_pull_reader object_value test")
+TEST_CASE("json_cursor object_value test")
 {
     std::string s = R"(
         {
@@ -299,7 +299,7 @@ TEST_CASE("json_pull_reader object_value test")
 
     std::istringstream is(s);
 
-    json_pull_reader reader(is);
+    json_cursor reader(is);
 
     REQUIRE_FALSE(reader.done());
     CHECK(reader.current().event_type() == staj_event_type::begin_object);